pyproject.toml
[tool.poetry]
name = "rest-svr"
version = "0.1.0"
description = ""
authors = ["jaeho kang <greennuri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {extras = ["all"], version = "^20.8"}
python-dotenv = "^1.0.1"
icecream = "^2.1.3"
langchain-community = "^0.0.19"
langchain = "^0.1.4"
elasticsearch = "^8.12.0"
elasticsearch-dsl = "^8.12.0"
eland = "^8.12.0"
tqdm = "^4.66.1"
huggingface = "^0.0.1"
huggingface-cli = "^0.1"
torch = "^2.1.2"
transformers = "^4.37.1"
sentence-transformers = "^2.2.2"
langchain-openai = "^0.0.5"
requests = "^2.31.0"
simsimd = "^3.7.7"
lark = "^1.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
